refactor: 코드 정렬 및 error 패키지 생성
