+++ src/main/java/lotto/domain/BonusNumber.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.ErrorMessage;
+import lotto.error.ErrorMessage;
 
 public class BonusNumber {
