+++ src/main/java/lotto/util/LottoGenerator.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public class LottoGenerator {
 
-    public static List<Integer> generateLotto(){
+    public static List<Integer> generateLotto() {
         return Randoms.pickUniqueNumbersInRange(1, 45, 6);
     }
 }
