+++ src/main/java/lotto/view/InputView.java
@@ -1,7 +1,7 @@
 package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
-import lotto.ErrorMessage;
+import lotto.error.ErrorMessage;
 
 public class InputView {
 
@@ -39,5 +39,4 @@ public class InputView {
             throw new IllegalArgumentException(ErrorMessage.EMPTY_INPUT.getValue());
         }
     }
-
 }
